# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuantSynthica Market API",
    "description": "Unified Financial-Data API integrating yfinance, TradingView-Screener, and Screener.in with quantitative analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/market/{symbol}/quote": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get real-time / latest stock quote",
        "operationId": "get_stock_quote_api_v1_market__symbol__quote_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/history": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get historical OHLCV data",
        "operationId": "get_stock_history_api_v1_market__symbol__history_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Data period e.g. 1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, max",
              "default": "1y",
              "title": "Period"
            },
            "description": "Data period e.g. 1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, max"
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Bar interval e.g. 1m, 2m, 5m, 15m, 30m, 60m, 90m, 1h, 1d, 5d, 1wk, 1mo",
              "default": "1d",
              "title": "Interval"
            },
            "description": "Bar interval e.g. 1m, 2m, 5m, 15m, 30m, 60m, 90m, 1h, 1d, 5d, 1wk, 1mo"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/dividends": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get dividend history",
        "operationId": "get_stock_dividends_api_v1_market__symbol__dividends_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DividendsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/splits": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get stock split history",
        "operationId": "get_stock_splits_api_v1_market__symbol__splits_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/recommendations": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get analyst consensus and price targets",
        "operationId": "get_analyst_recommendations_api_v1_market__symbol__recommendations_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalystRecommendationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/upgrades-downgrades": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get historical analyst rating changes",
        "operationId": "get_upgrades_downgrades_api_v1_market__symbol__upgrades_downgrades_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpgradesDowngradesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/options": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get options chain (calls and puts) by expiration date",
        "operationId": "get_options_chain_api_v1_market__symbol__options_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expiration date e.g. YYYY-MM-DD. If omitted, earliest expiration is returned",
              "title": "Date"
            },
            "description": "Expiration date e.g. YYYY-MM-DD. If omitted, earliest expiration is returned"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsChainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/news": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get latest corporate news and press releases",
        "operationId": "get_company_news_api_v1_market__symbol__news_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalystRecommendationsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "target_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriceTargetData"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "consensus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consensus"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "metadata"
        ],
        "title": "AnalystRecommendationsResponse"
      },
...
      "DividendItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "dividend": {
            "type": "number",
            "title": "Dividend"
          }
        },
        "type": "object",
        "required": [
          "date",
          "dividend"
        ],
        "title": "DividendItem"
      },
      "DividendsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DividendItem"
            },
            "type": "array",
            "title": "Data"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "count",
          "data",
          "metadata"
        ],
        "title": "DividendsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "interval": {
            "type": "string",
            "title": "Interval"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/OHLCVItem"
            },
            "type": "array",
            "title": "Data"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "period",
          "interval",
          "count",
          "data",
          "metadata"
        ],
        "title": "HistoryResponse"
      },
...
      "NewsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "news": {
            "items": {
              "$ref": "#/components/schemas/NewsItem"
            },
            "type": "array",
            "title": "News"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "count",
          "news",
          "metadata"
        ],
        "title": "NewsResponse"
      },
      "OHLCVItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "adj_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adj Close"
          },
          "volume": {
            "type": "integer",
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "date",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "OHLCVItem"
      },
...
      "OptionsChainResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "expiration_date": {
            "type": "string",
            "title": "Expiration Date"
          },
          "all_expirations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "All Expirations"
          },
          "calls": {
            "items": {
              "$ref": "#/components/schemas/OptionContractItem"
            },
            "type": "array",
            "title": "Calls"
          },
          "puts": {
            "items": {
              "$ref": "#/components/schemas/OptionContractItem"
            },
            "type": "array",
            "title": "Puts"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "expiration_date",
          "metadata"
        ],
        "title": "OptionsChainResponse"
      },
...
      "PriceTargetData": {
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean"
          },
          "median": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median"
          }
        },
        "type": "object",
        "title": "PriceTargetData"
      },
...
      "QuoteData": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency",
            "default": "INR"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price",
            "description": "Normalized current or last market price"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "previous_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change"
          },
          "change_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Percent"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "avg_volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Volume"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio"
          },
          "dividend_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Yield"
          },
          "fifty_two_week_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fifty Two Week High"
          },
          "fifty_two_week_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fifty Two Week Low"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "QuoteData"
      },
      "QuoteResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuoteData"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "data",
          "metadata"
        ],
        "title": "QuoteResponse"
      },
...
      "RecommendationItem": {
        "properties": {
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "strong_buy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strong Buy"
          },
          "buy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy"
          },
          "hold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold"
          },
          "sell": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sell"
          },
          "strong_sell": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strong Sell"
          }
        },
        "type": "object",
        "title": "RecommendationItem"
      },
      "ResponseMetadata": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Primary provider source, e.g. yfinance, tradingview, screener"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "Whether this response was served from cache",
            "default": false
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms",
            "description": "Time taken to fetch/compute"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "ResponseMetadata"
      },
...
      "SplitItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "split_ratio": {
            "type": "number",
            "title": "Split Ratio"
          }
        },
        "type": "object",
        "required": [
          "date",
          "split_ratio"
        ],
        "title": "SplitItem"
      },
      "SplitsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SplitItem"
            },
            "type": "array",
            "title": "Data"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "count",
          "data",
          "metadata"
        ],
        "title": "SplitsResponse"
      },
...
      "UpgradeDowngradeItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "firm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firm"
          },
          "to_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Grade"
          },
          "from_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Grade"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "UpgradeDowngradeItem"
      },
      "UpgradesDowngradesResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/UpgradeDowngradeItem"
            },
            "type": "array",
            "title": "Items"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "count",
          "metadata"
        ],
        "title": "UpgradesDowngradesResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}